Adds a Hashed MachineName enricher and extension
It does pretty much the same as the MachineName enrichment, but hashes the value so that the original value does not appear in logs. We can use it to create a correlation ID based on the machine name, to correlate all logs coming off of the same instalation, without being intrusive.